M2-#4: CSpace 扩容 32→64 slot/task (capabilities uint32→uint64)
微内核计划 M2 任务 4: 建立 CNode leaf,先支持 64-128 slot/任务。

实施 (扩到 64 slot,务实方案):
- tcb_t.capabilities: uint32 → uint64 (支持 64 bit 位图)
- tcb_t.cap_set[32] → cap_set[KERN_TASK_CAP_SLOTS] (配置驱动)
- BIT 宏: 1UL → 1ULL (32 位 ARM 上 unsigned long=32 位,BIT(32+) 原 UB)
- capability.c: cap_task_has/add/remove + cap_self_find_slot 的
  uint32_t bit → uint64_t bit (4 处)
- syscall.c: sys_task_create parent walk uint32_t bit → uint64_t (1 处)
- Kconfig: KERN_TASK_CAP_SLOTS default 32 → 64, range 8-64 不变
- 6 个 defconfig: KERN_TASK_CAP_SLOTS=32 → =64

128 slot 上限 (任务 4 上限) 留待 mint/copy 大量派生时再加 uint64[2]
两级 bitmap 或 CNode leaf 结构。当前 64 slot 已足够 (test_driver 等
最多用 ~10 cap/task)。

新测试 test_cap_cspace_over_32_slots:
- 填充 48 个 slot 跨过旧 32 边界
- 验证 capabilities 高 32 位被置位 (bit 32-47)
- 验证 slot 40 cap 可 resolve
- 注意: cap_task_add 只对 USER 任务登记,测试手工设 attrs=TASK_ATTR_USER

硬件验证 (RP2350 Pico2W, SMP=1):
- 2879/2879 All tests PASSED (5/5 跑 0 FAIL)
- bss 366252 → 370740 (+4.5KB: 24 task × 32 额外 slot × 2B + bitmap)

+++ src/kernel/cap/capability.c
@@ -114,7 +114,7 @@ static int cap_task_has(tcb_t *task, cap_id_t cap) {
     }
 
     for (int i = 0; i < CAP_TASK_CSPACE_SLOTS; i++) {
-        if ((task->capabilities & (uint32_t)BIT(i)) != 0 &&
+        if ((task->capabilities & (uint64_t)BIT(i)) != 0 &&
             task->cap_set[i] == cap) {
             return 1;
         }
@@ -137,7 +137,7 @@ static kern_err_t cap_task_add(tcb_t *task, cap_id_t cap) {
     }
 
     for (int i = 0; i < CAP_TASK_CSPACE_SLOTS; i++) {
-        uint32_t bit = (uint32_t)BIT(i);
+        uint64_t bit = (uint64_t)BIT(i);
         if ((task->capabilities & bit) == 0) {
             task->cap_set[i] = cap;
             task->capabilities |= bit;
@@ -154,7 +154,7 @@ static void cap_task_remove(tcb_t *task, cap_id_t cap) {
     }
 
     for (int i = 0; i < CAP_TASK_CSPACE_SLOTS; i++) {
-        uint32_t bit = (uint32_t)BIT(i);
+        uint64_t bit = (uint64_t)BIT(i);
         if ((task->capabilities & bit) != 0 && task->cap_set[i] == cap) {
             task->capabilities &= ~bit;
             task->cap_set[i] = 0;
@@ -771,7 +771,7 @@ cap_id_t cap_self_find_slot(tcb_t *owner, uint8_t obj_type, uint8_t index) {
     cap_id_t found = KERN_INVALID_ID;
 
     for (int i = 0; i < CAP_TASK_CSPACE_SLOTS; i++) {
-        uint32_t bit = (uint32_t)BIT(i);
+        uint64_t bit = (uint64_t)BIT(i);
         if ((owner->capabilities & bit) == 0) {
             continue;
         }
